+++ src/DrawableSuits/SuitTextureState.cs
@@ -7,6 +7,7 @@ internal sealed class SuitTextureState
     public int SuitId;
     public ulong OwnerClientId;
     public bool IsPlayerSpecific;
+    public bool IsLocalPlayerState;
     public string SuitName;
     public Material OriginalMaterial;
     public Texture OriginalTexture;